GDB-8757 save and revert ACL
## What
Implemented the save and revert ACL

## Why
* By requirement users should be able to save or rever changes in the ACL list.

## How
* Implemented the action for saving the ACL.
* Implemented the action for reverting the ACL.
* Implemented a new directive for capitalizing the input field value. This is used in the role property of the ACL rules.
* Implemented tests.

fixed test
