+++ src/app.js
@@ -8,6 +8,7 @@ import 'angular/core/interceptors/unauthorized.interceptor';
 import 'angular/core/directives/rdfresourcesearch/rdf-resource-search.directive';
 import 'angular/core/directives/languageselector/language-selector.directive';
 import 'angular/core/directives/angulartooltips/angular-tooltips.js';
+import 'angular/core/directives/uppercased.directive';
 
 // $translate.instant converts <b> from strings to &lt;b&gt
 // and $sce.trustAsHtml could not recognise that this is valid html
@@ -28,6 +29,7 @@ const modules = [
     'graphdb.framework.core.directives.rdfresourcesearch.rdfresourcesearch',
     'graphdb.framework.core.directives.languageselector.languageselector',
     'graphdb.framework.core.directives.angular-tooltips',
+    'graphdb.framework.core.directives.uppercased',
     'graphdb.framework.guides.services'
 ];
