+++ src/i18n/locale-en.json
@@ -155,16 +155,19 @@
                 "cancel_rule_editing": "Cancel rule editing",
                 "save_acl": "Save ACL",
                 "cancel_acl_saving": "Cancel"
-
             },
             "messages": {
                 "no_data": "There are no rules defined in current repository",
                 "invalid_form": "Fill in valid data in all fields before saving rule",
-                "delete_rule_confirmation": "Are you sure you want to delete the selected rule #{{index}}?"
+                "delete_rule_confirmation": "Are you sure you want to delete the selected rule #{{index}}?",
+                "revert_acl_list": "This operation will revert all the changes done in the ACL. Are you sure you want to proceed?",
+                "rules_updated": "ACL was updated successfully",
+                "rules_reverted": "ACL was reverted successfully"
             }
         },
         "errors": {
-            "loading_rules": "Error during ACL rules"
+            "loading_rules": "Error during ACL rules",
+            "updating_rules": "Error during ACL rules update"
         }
     },
     "config.name.label": "Config name",
