+++ src/js/angular/rest/aclmanagement.rest.service.js
@@ -10,15 +10,26 @@ const REPOSITORIES_ENDPOINT = 'rest/repositories';
 
 function AclManagementRestService($http) {
     return {
-        getRules,
-        updateRules
+        getAcl,
+        updateAcl
     };
 
-    function getRules(repositoryId) {
+    /**
+     * Fetches the ACL defined for given repository.
+     * @param {string} repositoryId The repository id for which the ACL to be fetched.
+     * @return {*}
+     */
+    function getAcl(repositoryId) {
         return $http.get(`${REPOSITORIES_ENDPOINT}/${repositoryId}/acl`);
     }
 
-    function updateRules(repositoryId, config) {
-        return $http.put(`${REPOSITORIES_ENDPOINT}/${repositoryId}/acl`, config);
+    /**
+     * Updates the ACL for provided repository by replacing the whole list with the provided one.
+     * @param {string} repositoryId The repository id for which the ACL to be updated.
+     * @param {ACListModel} data The ACListModel serialized to JSON.
+     * @return {*}
+     */
+    function updateAcl(repositoryId, data) {
+        return $http.put(`${REPOSITORIES_ENDPOINT}/${repositoryId}/acl`, data);
     }
 }
